Fix: Python example code were not tested
We've given users plenty of examples so they can try out different functions
on their own.
The problem is that we may have written some code that doesn't work at all.

Solution: use mktestdocs to test blocks of python code
I wrote a file located in /docs/test/ named python_test.py
The script will go to the destination path I've specified and test every block of code
that starts with ```python.
This also means that if I've written code in python but it's not intended to be executed, it's possible to put only ```py,
so that the code looks nice and well-formed but is not tested by mktestdocs.

To test locally, use the command below in /docs/test :
pytest python_test.py::test_run_python_code

Co-authored-by: Antonyjin <antony.jin@epitech.eu>