Update table with official runtime
and improve docs for runtimes in general

As a user I want to ensure that I use the proper runtime for programs, instances and confidentials, so we have to provide a kind of table with the official runtimes to use for each type of VM and their description.

We can place that table on the Docs page to be accessible to all users.